code review + react-bootstrap-table + modularity
