+++ caravel/assets/package.json
@@ -70,6 +70,7 @@
     "nvd3": "1.8.3",
     "react": "^15.2.0",
     "react-bootstrap": "^0.28.3",
+    "react-bootstrap-table": "^2.3.7",
     "react-dom": "^0.14.8",
     "react-grid-layout": "^0.12.3",
     "react-map-gl": "^1.0.0-beta-10",
